+++ config.ini
@@ -1,17 +1,42 @@
+# configuration data for the Peter project
+
 [DEFAULT]
+#Source is web = False
 Source is web = True
+
+# '%Y.%m.%d' = '2000.01.01'
+Date format = %%Y.%%m.%%d
 
 [Market Data Coupler]
 Market Data filename = market_data.csv
 Test Output Market Data filename = market_data_test_output.csv
-# '%Y.%m.%d'
-Date format = %%Y.%%m.%%d
-# '%H:%M:%S'
+
+# '%H:%M:%S' = '23:59:59'
 Time format = %%H:%%M:%%S 
 
 [TeleTrader]
 URLs filename = teletrader-links.csv
 TeleTrader to ISIN converter filename = ISIN.csv
+
 # source string example: 07.14./17:35
 # '%m.%d./%H:%M'
 Datetime format = %%m.%%d./%%H:%%M
+
+[Screening]
+Base Data filename = basedata.csv
+Short-list filename = shortlist.csv
+Slow SMA = 21
+Fast SMA = 7
+
+# suggested: False
+Show stocks above TAZ = False
+# suggested: True
+Show stocks in TAZ = True
+# suggested: True
+Show stocks below TAZ = True
+# suggested: False
+Show stocks with losses = False
+# suggested: False
+Show stocks with expired data = False
+# suggested: True
+Show stocks with position = True
