+++ screener.py
@@ -1,0 +1,13 @@
+'''
+Library to provide the screener class for the stocks.
+'''
+
+__last_change__ = '2017.08.30.'
+
+import sys
+
+def main():
+    pass
+
+if __name__ == '__main__':
+    sys.exit(main())
